refactor: move presentation files to component
